feat: order create

+++ model/product.go
@@ -6,7 +6,7 @@ type Product struct {
 	Info      string
 	AttrValue string
 	Price     float64
-	Total     uint64
+	Total     uint16
 }
 
 func (Product) TableName() string {
